+++ core/src/test/java/com/opensymphony/xwork2/conversion/impl/XWorkConverterTest.java
@@ -621,7 +621,7 @@ public class XWorkConverterTest extends XWorkTestCase {
         Class clazz = (Class) converter.convertValue(context, "java.util.Date", Class.class);
         assertEquals(Date.class.getName(), clazz.getName());
 
-        Class clazz2 = (Class) converter.convertValue(context, "com.opensymphony.xwork2.util.Bar", Class.class);
+        Class clazz2 = (Class) converter.convertValue(context, "com.opensymphony.xwork2.beans.util.Bar", Class.class);
         assertEquals(Bar.class.getName(), clazz2.getName());
 
         assertEquals(OgnlRuntime.NoConversionPossible, converter.convertValue(context, "com.opensymphony.xwork2.util.IDoNotExist", Class.class));
